summer/requirements.txt
# crewai
streamlit
plotly
fastapi
uvicorn
python-dotenv
langchain
redis
openai
google-generativeai
streamlit-lottie
